Some modification to OpenVanilla.h, so that now when you call OVBuffer::send or OVBuffer::update, you can specify the language in which the text is composed. This will be useful for applications that can adjust font/display behavior according to different langauge strings.
Beginning of writing OVXLoader. OVXTest has caused much pain and 
complain. Will be moving on. Also, as there is a new version of 
CarbonIM, in which some API has been changed, it will be necessary
in the next revisions to sync with this change. (Currently neither
OVXTest or OVXLoader has incorporated those changes in CarbonIM).