Added OVLoadableVersion() to OVLoadable.h
It's strongly recommended (and in most cases required) that you implement
this function in your loadable IM module. The reason is that OpenVanilla.h
may evolve. But as C++ programmers know, changes in class interface
result in ABSOLUTELY IMCOMPATIBLE SHARED LIBRARIES. This is due to the 
fact that C++ uses "virtual table" to dispatch class methods (for derived
classes). This is especially true when the number of parameters is changed
in any class method.

To prevent this, it's better for loadable IM to tell the loader which
version of OpenVanilla.h the IM is using. A well-designed loader should
not load any incompatible IM modules. Version information will be a good
tell.