Initial commit for experimental python interface for writing IM with OV.  The code shows how to mapping the C++ instances to Python instances within the SWIG framework. The python code can extend the C++ class and is called by the wrapping classes in c++.
The python wrap code is generated using SWIG 1.3.20 in the Fink distribution.
The code is tested in X.4.10 using the default python2.3.  The makefile will
need to be modified for testing in X.5 or using other versions of python.

The SWIG generated code is submitted as source for people who want to test it out 
without SWIG.

Also, the safty mechanism for loading only one copy of python has not be implemented
yet.  Namely, it might be problem if you use OVOFPythonBased and OVIMPythonBases
at the same time.


Build and install:

make OVIMPython
make DSTROOT=~ install