Checks in iVanilla SimpleBPMF and SimpleCJ.
iVanilla had been an "experiment" on iPhone/iPod Touch "Toolchain".
The experiment was done during the course of Oct-Nov 2007. Since
then the experiment had come to halt.

PLEASE READ THE README.txt IN Experiments/iVanilla/{SimpleBPMF,
SimpleCJ} BEFORE YOU PROCEED WITH THE CODE. THE CODE IS PROVIDED
AS IS. WE ARE NOT RESPONSIBLE FOR ANY POTENTIAL HARM. ALSO WE 
WON'T BE ABLE TO ANSWER ANY QUESTION ON IT.