An updated OVIMGeneric IM. This module will parse .cin into sqlite3 db automatically and also support reverse-lookup, associated phrase (with assoc.cin), user frequency learning.
For running on MacOS, OVCINSQList::load() in OVCINSQInfo has to be revised for adding code getting timestamp of .cin from filesystem.  The timestamp format on WIN32 is a 64-bit value representing the number of 100-nanosecond intervals since January 1, 1601 (UTC) put in two dwords. The timestamp is saved in the tablelist table of db, and checked everytime the IM is loaded.  If the .cin is updated, the old table will be drop and parse the new .cin into db.