Changes the behavior of using space as the first candidate key setting.
Before the commit, when the setting to use space as the first candidate
key is on, we just add space to the begin of the list of the candidates keys,
so the keys would be like "[SPACE]1234...", and the key 1 becomes the
key to select the second candidate.

Users expect that they can use both space and 1 to select the first
candidate. So, the keys should be like "[SPACE]234..".

This is for #30.